implicit casting fix
size of bmp file
image_size is a uint16_t, which would remove the need to static cast, but I believe if image_size is 4096. There's potential for overflow, again reintroducing the need to cast up to uint32_t. Not really sure what the code will do behind the scenes, so probably best to use the more readable width * height